Renamed factory mentions as it was not a fitting name.
